web/feat: projects cleaner refactoring
